segmentation/hsi-brain-segmentation/requirements.txt
cartopy==0.24.1
ipython==8.12.3
ipywidgets==8.1.3
matplotlib==3.7.5
mayavi==4.8.2
monai_weekly==1.4.dev2423
networkx==3.1
numpy==1.24.4
opencv_python==4.10.0.82
opencv_python_headless==4.9.0.80
pandas==2.0.3
Pillow==11.1.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.15.2
seaborn==0.13.2
segmentation_models_pytorch==0.3.3
spectral==0.23.1
torch==2.3.0
torchvision==0.18.0
wandb>=0.15.9